requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
attrs==24.2.0
binance==0.3
certifi==2024.7.4
charset-normalizer==3.3.2
dateparser==1.2.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
mysql-connector-python==9.0.0
numpy==2.0.1
pandas==2.2.2
pycryptodome==3.20.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
requests==2.32.3
six==1.16.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
websockets==12.0
yarl==1.9.4
